+++ packages/slack-utils/src/slack/utils/index.ts
@@ -2,3 +2,5 @@ export * from './unique';
 export * from './converter';
 export * from './trim';
 export * from './fetch';
+export * from './recursive';
+export * from './messages';