+++ packages/slack-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @kkkaoru/slack-utils
 
+## 1.4.0
+
+### Minor Changes
+
+- update slack utils
+
+## 1.3.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kkkaoru/openai-utils@1.0.2
+
 ## 1.3.12
 
 ### Patch Changes